# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/cms/media/ingestions": {
      "post": {
        "operationId": "CmsController_ingestMedia",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "name": "dataSourceId",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 2147483647,
              "type": "number"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Replaying a key with the same request returns the original result.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {},
                    "meta": {
                      "$ref": "#/components/schemas/ApiResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Cms"
        ]
      }
    },
    "/api/v1/admin/cms/media/ingestions/{ingestionId}/cleanup": {
      "post": {
        "operationId": "CmsController_retryMediaIngestionCleanup",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "name": "ingestionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {},
                    "meta": {
                      "$ref": "#/components/schemas/ApiResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Cms"
        ]
      }
    },
    "/api/v1/admin/cms/media": {
      "get": {
        "operationId": "CmsController_listMedia",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 2147483647,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "audio",
                "image",
                "pdf",
                "video"
              ]
            }
          },
          {
            "name": "processingStatus",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "processing",
                "ready",
                "failed",
                "quarantined"
              ]
            }
          },
          {
            "name": "lifecycle",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "archived"
              ]
            }
          },
          {
            "name": "dataSourceId",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 2147483647,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AdminMediaDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiPageResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Cms"
        ]
      }
    },
    "/api/v1/admin/cms/media/{mediaId}": {
      "get": {
        "operationId": "CmsController_getMedia",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/AdminMediaDetailDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Cms"
        ]
      }
    },
    "/api/v1/admin/cms/media/{mediaId}/quarantine": {
      "post": {
        "operationId": "CmsController_quarantineMedia",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/AdminMediaLifecycleResultDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Cms"
        ]
      }
    },
    "/api/v1/admin/cms/media/{mediaId}/archive": {
      "post": {
        "operationId": "CmsController_archiveMedia",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/AdminMediaLifecycleResultDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Cms"
        ]
      }
    },
    "/api/v1/admin/cms/exercises": {
      "get": {
        "operationId": "CmsController_listExercises",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 2147483647,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "lessonId",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 2147483647,
              "type": "number"
            }
          },
          {
            "name": "topicId",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 2147483647,
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "mcq",
                "fill_blank",
                "listening_choice",
                "arrange_sentence",
                "speaking_repeat"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "published",
                "archived"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AdminExerciseDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiPageResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Cms"
        ]
      },
      "post": {
        "operationId": "CmsController_createExercise",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExerciseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {},
                    "meta": {
                      "$ref": "#/components/schemas/ApiResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Cms"
        ]
      }
    },
    "/api/v1/admin/cms/exercises/{exerciseId}": {
      "get": {
        "operationId": "CmsController_getExercise",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "name": "exerciseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/AdminExerciseDetailDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Cms"
        ]
      }
    },
...
  },
  "info": {
    "title": "HSK System API",
    "description": "Generated from the backend code. Envelope, errors and request id: docs/api/api.md §1.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AdminMediaLifecycleResultDto": {
        "type": "object",
        "properties": {
          "idempotent": {
            "type": "boolean"
          },
          "media": {
            "$ref": "#/components/schemas/AdminMediaDto"
          }
        },
        "required": [
          "idempotent",
          "media"
        ]
      },
...
      "ApiResponseMetaDto": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "requestId": {
            "type": "string",
            "description": "Same value as the `X-Request-ID` response header."
          }
        },
        "required": [
          "timestamp",
          "requestId"
        ]
      },
...
      "ApiPageResponseMetaDto": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "requestId": {
            "type": "string",
            "description": "Same value as the `X-Request-ID` response header."
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "timestamp",
          "requestId",
          "pagination"
        ]
      },
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "details": {
            "type": "object",
            "additionalProperties": true
          },
          "code": {
            "type": "string",
            "description": "Stable machine-readable code, for example `UNAUTHORIZED`."
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ApiErrorEnvelopeDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorDto"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiResponseMetaDto"
          }
        },
        "required": [
          "success",
          "error",
          "meta"
        ]
      }
    },
    "parameters": {
      "RequestId": {
        "name": "X-Request-ID",
        "in": "header",
        "required": false,
        "description": "UUID or 32 hex characters; any other value is replaced by a new UUID. Echoed in the X-Request-ID response header and meta.requestId.",
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "ErrorEnvelope": {
        "description": "Error envelope.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
            }
          }
        }
      }
    }
  }
}